Stadium GUI, Club GUI, Stats GUI, Fixture chooser GUI done

+++ HCI-Project1-FootballLeague/HCI-Project1-FootballLeague/LeagueAdminTableWindow/LeagueAdminWindow.xaml.cs
@@ -24,6 +24,11 @@ namespace HCI_Project1_FootballLeague.LeagueAdminTableWindow
         public LeagueAdminWindow()
         {
             InitializeComponent();
+            PopulateData();
+        }
+
+        private void PopulateData()
+        {
             List<Administrator> admins = AdminDB.GetLeagueAdministrators();
             foreach (Administrator admin in admins)
             {
@@ -33,16 +38,39 @@ namespace HCI_Project1_FootballLeague.LeagueAdminTableWindow
 
         private void DeleteButton_Click(object sender, RoutedEventArgs e)
         {
-            var selectedAdmin = (Administrator)DataGridXAML.SelectedItem;
-            AdminDB.DeleteAdmin(selectedAdmin.AdminId);
+            if (DataGridXAML.SelectedItem != null)
+            {
+                var selectedAdmin = (Administrator)DataGridXAML.SelectedItem;
+                AdminDB.DeleteAdmin(selectedAdmin.AdminId);
+                DrawData();
+            } else
+            {
+                NotSelectedMessage();
+            }
+                
         }
 
         private void AddButton_Click(object sender, RoutedEventArgs e)
         {
             var userName = UserNameTB.Text;
             var password = PasswordTB.Text;
-            Administrator admin = new Administrator(userName, password, false);
-            AdminDB.AddAdmin(admin);
+            if(!"".Equals(userName) && !"".Equals(password))
+            {
+                Administrator admin = new Administrator(userName, password, false);
+                AdminDB.AddAdmin(admin);
+                DrawData();
+            }
+            else
+            {
+                MessageBox.Show("User name or password not entered!");
+            }
+            
+        }
+
+        private void DrawData()
+        {
+            DataGridXAML.Items.Clear();
+            PopulateData();
         }
 
         private void UpdateButton_Click(object sender, RoutedEventArgs e)
@@ -53,11 +81,21 @@ namespace HCI_Project1_FootballLeague.LeagueAdminTableWindow
                 UpdatePasswordLabel.Visibility = Visibility.Visible;
                 UserNameUpdate.Visibility = Visibility.Visible;
                 PasswordUpdate.Visibility = Visibility.Visible;
+                ConfirmUpdateButton.Visibility = Visibility.Visible;
                 var admin = (Administrator)DataGridXAML.SelectedItem;
                 UserNameUpdate.Text = admin.UserName;
                 PasswordUpdate.Text = admin.Password;
+            } 
+            else
+            {
+                NotSelectedMessage();
             }
-            
+
+        }
+
+        private static void NotSelectedMessage()
+        {
+            MessageBox.Show("Item not selected!");
         }
 
         private void ConfirmUpdateButton_Click(object sender, RoutedEventArgs e)
@@ -71,6 +109,16 @@ namespace HCI_Project1_FootballLeague.LeagueAdminTableWindow
                 bool isMain = false;
                 Administrator updatedAdmin = new Administrator(id, userName, password, isMain);
                 AdminDB.UpdateAdmin(updatedAdmin);
+                UpdateNameLabel.Visibility = Visibility.Hidden;
+                UpdatePasswordLabel.Visibility = Visibility.Hidden;
+                UserNameUpdate.Visibility = Visibility.Hidden;
+                PasswordUpdate.Visibility = Visibility.Hidden;
+                ConfirmUpdateButton.Visibility = Visibility.Hidden;
+                DrawData();
+            }
+            else
+            {
+                NotSelectedMessage();
             }
         }
     }
